修正用户更新API并更新用户仓库以支持长整型ID
在`UserController`中对`update`方法进行了修改，添加了`@Validated`注解以提高安全性。同时，`UserRepository`中的主键类型从`Integer`改为`Long`，以支持更大的用户ID。这些更改确保了用户更新操作的安全性和可扩展性。

+++ src/main/java/com/ovo307000/bigevent/repository/user/UserRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository("userUserRepository")
-public interface UserRepository extends JpaRepository<UserDTO, Integer>
+public interface UserRepository extends JpaRepository<UserDTO, Long>
 {
     List<UserDTO> findUsersByNickname(@Param("nickname") String nickname);
